+++ R/omega.R
@@ -14,7 +14,7 @@
 #'
 #' @examples 
 #' data(adj_karate)
-#' xi <- ComputeXi(adj_karate, FALSE, FALSE)
+#' xi <- compute_xi(adj_karate, FALSE, FALSE)
 #' FitOmega(adj_karate, xi, FALSE, FALSE)
 #' 
 FitOmega <- function(adj, xi, directed, selfloops){