+++ R/tests.R
@@ -40,7 +40,7 @@ conf.test <- function(graph, directed, selfloops, nempirical=NULL, parallel=NULL
   # if(!directed) xiregular <- xiregular + t(xiregular) - diag(diag(xiregular))
   xiregular <- ceiling(xiregular)
 
-  xiconfiguration <- ComputeXi(adj, directed, selfloops)
+  xiconfiguration <- compute_xi(adj, directed, selfloops)
 
   # if(nrow(adj)<50){
     loglikeregular <- extraDistr::dmvhyper(x = adj[ix], n = xiregular[ix], k = m, log = TRUE)
@@ -76,7 +76,7 @@ conf.test <- function(graph, directed, selfloops, nempirical=NULL, parallel=NULL
     n <- c(1, nrow(adj),ncol(adj))
     tmp <- vec2mat(adjr,directed,selfloops,n = n)
     if(!directed & n[2]==n[3]) tmp <- tmp + t(tmp) - diag(diag(tmp))
-    xiconfigurationr <- ComputeXi(tmp, directed, selfloops)
+    xiconfigurationr <- compute_xi(tmp, directed, selfloops)
     loglikeregularr <- extraDistr::dmvhyper(x = adjr, n = xiregular[ix], k = m, log = TRUE)
     loglikeconfr <- extraDistr::dmvhyper(x = adjr, n = xiconfigurationr[ix], k = m, log = TRUE)
     -2 * (loglikeregularr-loglikeconfr)